requirements.txt
anthropic>=0.40.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
functions-framework>=3.5.0
python-dotenv>=1.0.0
pytz>=2024.1
Pillow>=10.0.0
pytest>=8.0.0
